+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
 
+## Unreleased
+
+### Added
+
+- Fluent hook for logrus
+
+
 ## 0.1.0 - 2017-06-19
 
 - Initial release